requirements.txt
beautifulsoup4>=4.12
patchright>=1.52.5
playwright>=1.52
pypdf>=5.0
python-dotenv>=1.0
requests>=2.31